calendar

+++ src/main/java/sg/edu/iss/team6/model/Student.java
@@ -1,6 +1,6 @@
 package sg.edu.iss.team6.model;
 import javax.persistence.*;
-import javax.validation.constraints.Size;
+
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.Size;
@@ -40,7 +40,7 @@ public class Student {
     @Size(min=2, max=50, message="Address must be 2-50 characters long")
     private String address;
 
-    private long gpa;
+    private double gpa;
 
     @OneToMany(mappedBy="student")
     @JsonIgnore
